DDK v2.1 related checkin #2
Where is the DDK v2.1 related checkin #1? Actually, the previous one
commit referred as DDK v2.0 checkin #3, is in fact the DDK v2.1 related
checkin #1. And this checkin is the final checkin, as sent for
fabrication of DDK v2.1. It contains the following changes from v2.0:
+ Schematic:
	> Pulled-up the D of shift resister to get fixed default off state
		of the LED array
	> Renamed the G(ermanium) Prot(ection) to R(everse) Protection, as
		the normal diode seems more promising than the germanium diode,
		at least in the forward bias mode
	> Version change to 2.1
+ Layout:
	> Did font size adjustments to make "products@..." clearer
	> Pushed the DB9 connector more inside the board
		. Change the footprint of the DB9 connector, with full outline
		. And accordingly modified the required track layings
	> Replaced the OSHW logo outline by a filled one
	> Spaced away the GND header from the Vcc header
	> Done the placement of the added pull-up resistor
	> Reloaded the netlist for the updated diode label, as per the
		schematic change
	> Moved the IR LED a bit inside to avoid too much protruding out, as
		it may get broken bcoz of that
	> And finally, the version change to v2.1

As changes has been there from the beginning, all the associated output
files have been regenerated. Additionally, gerber files also have been
generated, as required for the fabrication, this time.

Created separate folders for links to the pdfs & gerbers sent out for
fabrication, and accordingly updated the README.md with that info.